refactor: restructure src directory for better organization
