+++ src/index.ts
@@ -1,25 +1,34 @@
 // Core BhaSha engine
-export { BhaSha } from './bhasha-engine';
+export { BhaSha } from './native/bhasha-engine';
 
 // Engine architecture
-export { EngineFactory } from './engines';
-export type { LanguageEngine } from './engines';
-export { BaseLanguageEngine } from './engines';
-export { GujaratiEngine } from './engines';
-export { HindiEngine } from './engines';
-export { EnglishEngine } from './engines';
+export { EngineFactory } from './native/engines';
+export type { LanguageEngine } from './native/engines';
+export { BaseLanguageEngine } from './native/engines';
+export { GujaratiEngine } from './native/engines';
+export { HindiEngine } from './native/engines';
+export { EnglishEngine } from './native/engines';
 
 // React Hook
-export { useBhaShaIME } from './useBhaShaIME';
-export type { UseBhaShaIMEOptions, UseBhaShaIMEReturn } from './useBhaShaIME';
-export { useDirectInputTransliteration } from './useDirectInputTransliteration';
+export { useBhaShaIME } from './hooks/useBhaShaIME';
+export type {
+  UseBhaShaIMEOptions,
+  UseBhaShaIMEReturn,
+} from './hooks/useBhaShaIME';
+export { useDirectInputTransliteration } from './hooks/useDirectInputTransliteration';
 
 // React Components
-export { BhaShaInput } from './BhaShaInput';
-export type { BhaShaInputProps, BhaShaInputRef } from './BhaShaInput';
+export { BhaShaInput } from './components/BhaShaInput';
+export type {
+  BhaShaInputProps,
+  BhaShaInputRef,
+} from './components/BhaShaInput';
 
-export { BhaShaTextarea } from './BhaShaTextarea';
-export type { BhaShaTextareaProps, BhaShaTextareaRef } from './BhaShaTextarea';
+export { BhaShaTextarea } from './components/BhaShaTextarea';
+export type {
+  BhaShaTextareaProps,
+  BhaShaTextareaRef,
+} from './components/BhaShaTextarea';
 
 // Language types and utilities
 export {
@@ -35,10 +44,10 @@ export type {
 } from './types';
 
 // Default export for convenience
-import { BhaSha } from './bhasha-engine';
-import { useBhaShaIME } from './useBhaShaIME';
-import { BhaShaInput } from './BhaShaInput';
-import { BhaShaTextarea } from './BhaShaTextarea';
+import { BhaSha } from './native/bhasha-engine';
+import { useBhaShaIME } from './hooks/useBhaShaIME';
+import { BhaShaInput } from './components/BhaShaInput';
+import { BhaShaTextarea } from './components/BhaShaTextarea';
 
 export default {
   BhaSha,
